pyproject.toml
[tool.poetry]
name = "mathproblemcreator"
version = "0.1.0"
description = ""
authors = ["76img <76img>"]

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
